test: minor changes to finch cp e2e tests
